pyproject.toml
[tool.poetry]
name = "hsa-homework-3"
version = "0.1.0"
description = ""
authors = ["BohdanPenkov <bohdanpenkov@planeks.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
schedule = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
